+++ src/main/java/seedu/duke/Exceptions/DukeError.java
@@ -1,4 +1,4 @@
-package seedu.duke.errors;
+package seedu.duke.Exceptions;
 
 public class DukeError extends Exception{
     public DukeError(String message) {
