+++ src/main/java/seedu/duke/commands/Command.java
@@ -1,6 +1,6 @@
 package seedu.duke.commands;
 
-import seedu.duke.errors.DukeError;
+import seedu.duke.Exception.DukeError;
 import seedu.duke.exercisegenerator.GenerateExercise;
 import seedu.duke.ui.Ui;
