+++ src/test/java/seedu/duke/exercisegenerator/GenerateDifficultyExerciseTest.java
@@ -1,7 +1,7 @@
 package seedu.duke.exercisegenerator;
 
 import org.junit.jupiter.api.Test;
-import seedu.duke.errors.DukeError;
+import seedu.duke.Exception.DukeError;
 import seedu.duke.exersisedata.ExerciseData;
 
 import java.util.ArrayList;
